*AND + w3 workflow works with springer

+++ disambiguation/src/main/java/pl/edu/icm/coansys/disambiguation/clustering/ClusterElement.java
@@ -27,10 +27,10 @@ package pl.edu.icm.coansys.disambiguation.clustering;
  */
 
 public class ClusterElement implements Comparable<Object>{
-	private double sim;
+	private float sim;
 	private int index;
 	
-	public ClusterElement(double sim, int index){
+	public ClusterElement(float sim, int index){
 		this.sim=sim;
 		this.index=index;
 	}
@@ -43,11 +43,11 @@ public class ClusterElement implements Comparable<Object>{
             this.index = index;
         }
         
-        public double getSim() {
+        public float getSim() {
             return sim;
         }
 
-        public void setSim(double sim) {
+        public void setSim(float sim) {
             this.sim = sim;
         }
         
@@ -60,7 +60,7 @@ public class ClusterElement implements Comparable<Object>{
                     throw new ClassCastException("" +
 				"Comparison between "+this.getClass()+" and "+o2.getClass()+" is illegal!");
                 }
-		double count = this.sim-((ClusterElement)o2).sim; 
+		float count = this.sim-((ClusterElement)o2).sim; 
 		if(count>0) {
                     return 1;
                 }
